lint fix, automatic eslinting & refactoring
