+++ nodemon.json
@@ -1,5 +1,5 @@
 {
-    "exec": "npm run start",
+    "exec": "npm run lint && npm run start",
     "watch": ["src/*"],
     "ext": "js, json"
 }
